Add PR workflow, improve config error message (#1)
* Improves AWS config validation error message

Enhances the AWS config validation error message to provide more specific guidance to the user when an AWS SSO session has expired or they are not logged in. This makes it easier for the user to understand the problem and how to fix it.

* Adds pull request workflow

Sets up a GitHub Actions workflow to automatically check code formatting, lint, run tests, and build the project on every pull request to the `master` branch.

This ensures code quality and consistency by catching potential issues early in the development process.

* style: formatting

* Improves CI workflow for faster builds

Optimizes CI workflows by adding rust-cache and setting environment variables to improve build times.
This change also enforces usage of locked dependencies.

* Renames pull request workflow file

Updates the pull request workflow filename to adhere to naming conventions.

* Adds dependabot configuration file

Adds a dependabot configuration file to automatically
check and update cargo dependencies on a weekly basis.

* fix: typo in build command

* ci: also use --locked on clippy